notifications (not real time)

+++ src/app/api/posts/[postId]/likes/route.ts
@@ -59,44 +59,45 @@ export async function POST(
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
 
-    // const post = await prisma.post.findUnique({
-    //   where: { id: postId },
-    //   select: {
-    //     userId: true,
-    //   },
-    // });
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true,
+      },
+    });
 
-    // if (!post) {
-    //   return Response.json({ error: "Post not found" }, { status: 404 });
-    // }
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
+    }
 
-    // await prisma.$transaction([
-    await prisma.like.upsert({
-      where: {
-        userId_postId: {
+    await prisma.$transaction([
+      prisma.like.upsert({
+        where: {
+          userId_postId: {
+            userId: loggedInUser.id,
+            postId,
+          },
+        },
+        create: {
           userId: loggedInUser.id,
           postId,
         },
-      },
-      create: {
-        userId: loggedInUser.id,
-        postId,
-      },
-      update: {},
-    });
-    //   ...(loggedInUser.id !== post.userId
-    //     ? [
-    //         prisma.notification.create({
-    //           data: {
-    //             issuerId: loggedInUser.id,
-    //             recipientId: post.userId,
-    //             postId,
-    //             type: "LIKE",
-    //           },
-    //         }),
-    //       ]
-    //     : []),
-    // ]);
+        update: {},
+      }),
+      // nếu người đang đăng nhập ko phải là người đăng bài thì mới thông báo
+      ...(loggedInUser.id !== post.userId
+        ? [
+            prisma.notification.create({
+              data: {
+                issuerId: loggedInUser.id,
+                recipientId: post.userId,
+                postId,
+                type: "LIKE",
+              },
+            }),
+          ]
+        : []),
+    ]);
 
     return new Response();
   } catch (error) {
@@ -116,33 +117,33 @@ export async function DELETE(
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
 
-    // const post = await prisma.post.findUnique({
-    //   where: { id: postId },
-    //   select: {
-    //     userId: true,
-    //   },
-    // });
-
-    // if (!post) {
-    //   return Response.json({ error: "Post not found" }, { status: 404 });
-    // }
-
-    // await prisma.$transaction([
-    await prisma.like.deleteMany({
-      where: {
-        userId: loggedInUser.id,
-        postId,
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true,
       },
     });
-    //   prisma.notification.deleteMany({
-    //     where: {
-    //       issuerId: loggedInUser.id,
-    //       recipientId: post.userId,
-    //       postId,
-    //       type: "LIKE",
-    //     },
-    //   }),
-    // ]);
+
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
+    }
+
+    await prisma.$transaction([
+      prisma.like.deleteMany({
+        where: {
+          userId: loggedInUser.id,
+          postId,
+        },
+      }),
+      prisma.notification.deleteMany({
+        where: {
+          issuerId: loggedInUser.id,
+          recipientId: post.userId,
+          postId,
+          type: "LIKE",
+        },
+      }),
+    ]);
 
     return new Response();
   } catch (error) {
